Created a basic logger
Makes debugging the MCP exceptions easier
