+++ McpServer.cs
@@ -10,31 +10,6 @@ using Microsoft.Extensions.Options;
 
 namespace RhMcp;
 
-internal sealed class RhinoLoggerProvider : ILoggerProvider
-{
-    public ILogger CreateLogger(string categoryName) => new RhinoLogger(categoryName);
-    public void Dispose() { }
-
-    private sealed class RhinoLogger : ILogger
-    {
-        private readonly string _category;
-        public RhinoLogger(string category) => _category = category;
-
-        public IDisposable? BeginScope<TState>(TState state) where TState : notnull => null;
-        public bool IsEnabled(LogLevel logLevel) => logLevel >= LogLevel.Information;
-
-        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state,
-            Exception? exception, Func<TState, Exception?, string> formatter)
-        {
-            if (!IsEnabled(logLevel)) return;
-            var msg = formatter(state, exception);
-            RhinoApp.WriteLine($"[Rhino MCP][{logLevel}] {_category}: {msg}");
-            if (exception is not null)
-                RhinoApp.WriteLine($"[Rhino MCP]   {exception.GetType().Name}: {exception.Message}\n{exception.StackTrace}");
-        }
-    }
-}
-
 internal sealed class McpServer : IDisposable
 {
     private WebApplication? _app;
@@ -52,6 +27,8 @@ internal sealed class McpServer : IDisposable
         {
             var builder = WebApplication.CreateSlimBuilder();
             builder.Logging.ClearProviders();
+            builder.Logging.AddProvider(new RhinoLoggerProvider());
+            builder.Logging.SetMinimumLevel(LogLevel.Information);
             builder.Services.Configure<KestrelServerOptions>(o => o.ListenLocalhost(port));
 
             builder.Services.AddSingleton(doc);
